+++ packages/tsdk/src/format.ts
@@ -1,8 +1,0 @@
-// import prettier from "prettier";
-
-export function formatTS(content: string) {
-  // return prettier.format(content, {
-  //   parser: "typescript",
-  // });
-  return content;
-}