+++ packages/tsdk/src/prettier.ts
@@ -1,0 +1,21 @@
+import { execSync } from 'child_process';
+import path from 'path';
+
+import { config, packageFolder } from './config';
+
+export async function runPrettier() {
+  const rootDir = path.resolve(process.cwd(), config.monorepoRoot || './');
+  const formatDir = path.resolve(process.cwd(), config.packageDir, packageFolder);
+
+  try {
+    // prettier 3.x
+    execSync(`node ${rootDir}/node_modules/prettier/bin/prettier.cjs ${formatDir} --write`);
+  } catch (e) {
+    try {
+      // prettier 2.x
+      execSync(`node ${rootDir}/node_modules/prettier/bin-prettier ${formatDir} --write`);
+    } catch (e) {
+      //
+    }
+  }
+}